pixelssl/requirements.txt
pyyaml
numpy
scipy
Pillow
opencv-python
